+++ examples/src/todomvc-experimental/src/livestore/schema.ts
@@ -8,7 +8,7 @@ const todos = DbSchema.table('todos', {
   id: DbSchema.text({ primaryKey: true }),
   text: DbSchema.text({ default: '' }),
   completed: DbSchema.boolean({ default: false }),
-  deleted: DbSchema.integer({ nullable: true }),
+  deleted: DbSchema.integer({ nullable: true, schema: Schema.DateFromNumber }),
 })
 
 const app = DbSchema.table(