+++ packages/@livestore/common/src/schema/EventId.test.ts
@@ -5,8 +5,8 @@ import { EventId } from './mod.js'
 
 Vitest.describe('EventId', () => {
   Vitest.test('nextPair', () => {
-    const e_0_0 = EventId.make({ global: 0, local: 0 })
-    expect(EventId.nextPair(e_0_0, false).id).toStrictEqual({ global: 1, local: 0 })
-    expect(EventId.nextPair(e_0_0, true).id).toStrictEqual({ global: 0, local: 1 })
+    const e_0_0 = EventId.make({ global: 0, client: 0 })
+    expect(EventId.nextPair(e_0_0, false).id).toStrictEqual({ global: 1, client: 0 })
+    expect(EventId.nextPair(e_0_0, true).id).toStrictEqual({ global: 0, client: 1 })
   })
 })