+++ packages/@livestore/web/src/web-worker/client-session/trim-batch.test.ts
@@ -7,10 +7,10 @@ import { trimPushBatch } from './trim-batch.js'
 describe('trimPushBatch', () => {
   it('should return same batch', () => {
     const batch = [
-      { id: EventId.make({ global: 0, local: 1 }), parentId: EventId.make({ global: 0, local: 0 }) },
-      { id: EventId.make({ global: 0, local: 2 }), parentId: EventId.make({ global: 0, local: 1 }) },
-      { id: EventId.make({ global: 1, local: 0 }), parentId: EventId.make({ global: 0, local: 0 }) },
-      { id: EventId.make({ global: 1, local: 1 }), parentId: EventId.make({ global: 1, local: 0 }) },
+      { id: EventId.make({ global: 0, client: 1 }), parentId: EventId.make({ global: 0, client: 0 }) },
+      { id: EventId.make({ global: 0, client: 2 }), parentId: EventId.make({ global: 0, client: 1 }) },
+      { id: EventId.make({ global: 1, client: 0 }), parentId: EventId.make({ global: 0, client: 0 }) },
+      { id: EventId.make({ global: 1, client: 1 }), parentId: EventId.make({ global: 1, client: 0 }) },
     ] as MutationEvent.AnyEncoded[]
 
     const trimmed = trimPushBatch(batch)
@@ -20,13 +20,13 @@ describe('trimPushBatch', () => {
 
   it('should trim the batch', () => {
     const batch = [
-      { id: EventId.make({ global: 0, local: 1 }), parentId: EventId.make({ global: 0, local: 0 }) },
-      { id: EventId.make({ global: 0, local: 2 }), parentId: EventId.make({ global: 0, local: 1 }) },
+      { id: EventId.make({ global: 0, client: 1 }), parentId: EventId.make({ global: 0, client: 0 }) },
+      { id: EventId.make({ global: 0, client: 2 }), parentId: EventId.make({ global: 0, client: 1 }) },
       // should trim above
-      { id: EventId.make({ global: 0, local: 1 }), parentId: EventId.make({ global: 0, local: 0 }) },
-      { id: EventId.make({ global: 0, local: 2 }), parentId: EventId.make({ global: 0, local: 1 }) },
-      { id: EventId.make({ global: 1, local: 0 }), parentId: EventId.make({ global: 0, local: 0 }) },
-      { id: EventId.make({ global: 1, local: 1 }), parentId: EventId.make({ global: 1, local: 0 }) },
+      { id: EventId.make({ global: 0, client: 1 }), parentId: EventId.make({ global: 0, client: 0 }) },
+      { id: EventId.make({ global: 0, client: 2 }), parentId: EventId.make({ global: 0, client: 1 }) },
+      { id: EventId.make({ global: 1, client: 0 }), parentId: EventId.make({ global: 0, client: 0 }) },
+      { id: EventId.make({ global: 1, client: 1 }), parentId: EventId.make({ global: 1, client: 0 }) },
     ] as MutationEvent.AnyEncoded[]
 
     const trimmed = trimPushBatch(batch)
@@ -36,9 +36,9 @@ describe('trimPushBatch', () => {
 
   it('should trim the batch', () => {
     const batch = [
-      { id: EventId.make({ global: 0, local: 1 }), parentId: EventId.make({ global: 0, local: 0 }) },
+      { id: EventId.make({ global: 0, client: 1 }), parentId: EventId.make({ global: 0, client: 0 }) },
       // should trim above
-      { id: EventId.make({ global: 0, local: 1 }), parentId: EventId.make({ global: 0, local: 0 }) },
+      { id: EventId.make({ global: 0, client: 1 }), parentId: EventId.make({ global: 0, client: 0 }) },
     ] as MutationEvent.AnyEncoded[]
 
     const trimmed = trimPushBatch(batch)